backend/requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
requests
google-adk
google-cloud-aiplatform
pydantic
